requirements.txt
decorator==4.0.10
Django==1.10.4
django-crontab==0.7.1
ipython-genutils==0.2.0
numpy==1.13.0
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.9
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.1.3
python-bitcoinlib==0.10.1
python-bitcoinrpc==1.0
pytz==2016.10
simplegeneric==0.8.1
six==1.10.0
style==1.1.0
traitlets==4.3.1
update==0.0.1
uWSGI==2.0.15
wcwidth==0.1.7
